skip test on windows

+++ src/TaskRunner.test.js
@@ -4,6 +4,8 @@ const nodeCleanup = require('node-cleanup');
 const GuaranteedTask = require('./GuaranteedTask');
 const TaskRunner = require('./TaskRunner');
 
+const isWin = process.platform === 'win32';
+
 // prevent exit once for mocking
 let stoppedOnce = false;
 nodeCleanup((exitCode, signal) => {
@@ -175,31 +177,33 @@ describe('TaskRunner', () => {
     });
   });
 
-  it('should shut down gracefully', async () => {
-    const taskRunner = new TaskRunner(
-      { Task: LongRunningTask },
-    );
-    await taskRunner.start();
-    taskRunner.addTask(1000);
-    process.kill(process.pid, 'SIGINT');
-    expect(mock).toBeCalledTimes(1);
-    const waitForStop = () => new Promise((resolve) => {
-      const isStopped = () => !taskRunner.running && !taskRunner.stopping;
-      const doCheck = () => {
-        setTimeout(() => {
-          if (isStopped()) {
-            resolve();
-          } else {
-            doCheck();
-          }
-        }, 100);
-      };
-      doCheck();
-    });
-    await waitForStop();
-    expect(taskRunner.running).toBeFalsy();
+  if (!isWin) {
+    it('should shut down gracefully', async () => {
+      const taskRunner = new TaskRunner(
+        { Task: LongRunningTask },
+      );
+      await taskRunner.start();
+      taskRunner.addTask(1000);
+      process.kill(process.pid, 'SIGINT');
+      expect(mock).toBeCalledTimes(1);
+      const waitForStop = () => new Promise((resolve) => {
+        const isStopped = () => !taskRunner.running && !taskRunner.stopping;
+        const doCheck = () => {
+          setTimeout(() => {
+            if (isStopped()) {
+              resolve();
+            } else {
+              doCheck();
+            }
+          }, 100);
+        };
+        doCheck();
+      });
+      await waitForStop();
+      expect(taskRunner.running).toBeFalsy();
 
-    // cleanup
-    taskRunner.closeDb();
-  });
+      // cleanup
+      taskRunner.closeDb();
+    });
+  }
 });
